backend_認可機能の追加と削除権限の確認機能の実装 #9

+++ src/pages/api/item/delete/[id].js
@@ -1,10 +1,17 @@
+import auth from 'utils/auth';
 import connectDB from 'utils/database';
 import { ItemModel } from 'utils/schemaModels';
 
 const deleteItem = async (req, res) => {
   try {
-    const _id = req.query.id;
     await connectDB();
+    
+    const _id = req.query.id;
+    const email = req.body.email;
+
+    const item = await ItemModel.findById(_id);
+    if (email !== item.email) throw new Error();
+
     await ItemModel.deleteOne({ _id });
     return res.status(200).json({ message: 'アイテムを削除しました' });
   } catch (error) {
@@ -12,4 +19,4 @@ const deleteItem = async (req, res) => {
   }
 };
 
-export default deleteItem;
+export default auth(deleteItem);
